Prevent warnings during upgrades
The `maintenance upgrade` script kills the js-controller, so there is nothing left to kill during shutdown and a warning would printed:

/opt/scripts/iobroker_startup.sh: line 497: kill: `': not a pid or valid job spec

+++ debian/scripts/iobroker_startup.sh
@@ -499,7 +499,7 @@ shut_down() {
   echo "Recived termination signal (SIGTERM)."
   echo "Shutting down ioBroker..."
   pid=$(ps -ef | awk '/[j]s.controller/{print $2}')
-  kill -SIGTERM "$pid"
+  [[ -n "$pid" ]] && kill -SIGTERM "$pid"
   exit
 }
